CMS_auth/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
WTForms==3.1.2
alembic==1.13.1
blinker==1.7.0
boto3==1.34.48
botocore==1.34.48
click==8.1.7
email_validator==2.1.0.post1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.2
MarkupSafe==2.1.5
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
